modified duration data type

+++ src/timer.h
@@ -1,3 +1,3 @@
-void wait(int duration){
+void wait(size_t duration){
 
 }
